rename ember-mocha to ember-mocha2
