+++ addon-test-support/ember-mocha2/index.js
@@ -3,9 +3,9 @@
 export { loadTests } from './test-loader';
 
 import { loadTests } from './test-loader';
-import setupTest from 'ember-mocha/setup-test';
-import setupRenderingTest from 'ember-mocha/setup-rendering-test';
-import setupApplicationTest from 'ember-mocha/setup-application-test';
+import setupTest from 'ember-mocha2/setup-test';
+import setupRenderingTest from 'ember-mocha2/setup-rendering-test';
+import setupApplicationTest from 'ember-mocha2/setup-application-test';
 import { beforeEach, afterEach } from 'mocha';
 import { setResolver, resetOnerror } from '@ember/test-helpers';
 import Ember from 'ember';
