+++ tests/test-helper.js
@@ -1,4 +1,4 @@
-import { setResolver, start } from 'ember-mocha';
+import { setResolver, start } from 'ember-mocha2';
 import resolver from './helpers/resolver';
 
 setResolver(resolver);
